esp-01 uint16_t

+++ ImpCounter85/src/sleep_counter.h
@@ -19,7 +19,7 @@
   - 
 */
 
-void gotoDeepSleep( unsigned long seconds, unsigned long *counter, unsigned long *counter2);
+void gotoDeepSleep( uint16_t seconds, uint16_t *counter, uint16_t *counter2);
 void resetWatchdog();
 
 #endif
