+++ ImpCounterESP/src/MasterI2C.h
@@ -7,15 +7,15 @@
 
 
 struct SlaveStats {
-	uint32_t bytesReady;
-	uint32_t masterWakeEvery;
-	uint32_t measurementEvery;
-	uint32_t vcc;
+	uint16_t bytesReady;
+	uint16_t masterWakeEvery;
+	uint16_t measurementEvery;
+	uint16_t vcc;
 	uint8_t bytesPerMeasurement;
 	uint8_t version;
 	uint8_t numberOfSensors;
 	uint8_t dummy;
-	uint32_t deviceID;
+	uint16_t deviceID;
 }; //should be *16bit https://github.com/esp8266/Arduino/issues/1825
 
 
@@ -24,13 +24,13 @@ class MasterI2C
  protected:
 	 void gotoFirstByte();
 	 byte getNextByte();
-	 uint32_t getUint();
+	 uint16_t getUint();
 	 uint8_t getByte();
 
  public:
 	 void begin();
 	 void sendCmd( const char cmd );
-	 uint16_t getSlaveStorage( byte* storage, const uint16_t maxStorageSize, const uint32_t bytesToFetch );
+	 uint16_t getSlaveStorage( byte* storage, const uint16_t maxStorageSize, const uint16_t bytesToFetch );
 	 SlaveStats getSlaveStats();
 };
